Add SPDX License Identifier to Multicall2.sol
It fixes the compiler warning `license identifier not provided in source file` and is overall good practice to add the license identifier on a per-file basis as in the software industry no license specified defaults to all rights reserved, which could potentially prevent anyone to have the rights to deploy the individual Multicall2.sol. Though, this is technically not necessary in the particular case of MIT and Unlicense licenses.